change name form lines to split and match directories
