+++ internal/split/iter.go
