+++ pkg/ext/ext.go
@@ -2,8 +2,8 @@ package ext
 
 import (
 	"github.com/augmentable-dev/flite/internal/http"
-	"github.com/augmentable-dev/flite/internal/lines"
 	"github.com/augmentable-dev/flite/internal/readfile"
+	"github.com/augmentable-dev/flite/internal/split"
 	"github.com/augmentable-dev/flite/internal/yaml"
 	_ "github.com/mattn/go-sqlite3"
 	"go.riyazali.net/sqlite"
@@ -11,7 +11,7 @@ import (
 
 func init() {
 	sqlite.Register(func(api *sqlite.ExtensionApi) (sqlite.ErrorCode, error) {
-		if err := api.CreateModule("lines", lines.NewVTab(),
+		if err := api.CreateModule("split", split.NewVTab(),
 			sqlite.EponymousOnly(true), sqlite.ReadOnly(true)); err != nil {
 			return sqlite.SQLITE_ERROR, err
 		}
